refactor: Rename namespaces for Create and Update Blogger commands
